Handle double exam total obtained

+++ SMS/ExamsManagement/ExamDataEntry/ExamDataEntryPage.xaml.cs
@@ -752,7 +752,7 @@ namespace SMS.ExamsManagement.ExamDataEntry
                         else
                         {
                             subject_total = Convert.ToDouble(ede.subject_total);    
-                            ede.subject_total_int = Convert.ToInt32(ede.subject_total);
+                            ede.subject_total_int = Convert.ToDouble(ede.subject_total);
                             total_marks = total_marks + subject_total;
                         }
 
@@ -770,7 +770,7 @@ namespace SMS.ExamsManagement.ExamDataEntry
                         else
                         {
                             subject_total = Convert.ToDouble(ede.subject_total);
-                            ede.subject_total_int = Convert.ToInt32(ede.subject_total);
+                            ede.subject_total_int = Convert.ToDouble(ede.subject_total);
                             total_marks = total_marks + subject_total;
                         }
                     }
@@ -787,7 +787,7 @@ namespace SMS.ExamsManagement.ExamDataEntry
                         else
                         {
                             subject_total = Convert.ToDouble(ede.subject_total);
-                            ede.subject_total_int = Convert.ToInt32(ede.subject_total);
+                            ede.subject_total_int = Convert.ToDouble(ede.subject_total);
                             total_marks = total_marks + subject_total;
                         }
                     }
@@ -925,8 +925,8 @@ namespace SMS.ExamsManagement.ExamDataEntry
                                 cmd.Parameters.Add("@subject_percentage", MySqlDbType.VarChar).Value = ede_s.subject_percentage;
                                 cmd.Parameters.Add("@subject_total", MySqlDbType.VarChar).Value = ede_s.subject_total;
                                 cmd.Parameters.Add("@subject_obtained", MySqlDbType.VarChar).Value = ede_s.subject_obtained;
-                                cmd.Parameters.Add("@subject_total_int", MySqlDbType.Int32).Value = ede_s.subject_total_int;
-                                cmd.Parameters.Add("@subject_obtained_int", MySqlDbType.Int32).Value = ede_s.subject_obtained_int;
+                                cmd.Parameters.Add("@subject_total_int", MySqlDbType.Double).Value = ede_s.subject_total_int;
+                                cmd.Parameters.Add("@subject_obtained_int", MySqlDbType.Double).Value = ede_s.subject_obtained_int;
 
                                 cmd.Parameters.Add("@total_marks", MySqlDbType.VarChar).Value = ede_s.total_marks;
                                 cmd.Parameters.Add("@total_remarks", MySqlDbType.VarChar).Value = ede_s.total_remarks;
